Move PickItemMenu.Search to its own class TextSearch
